Revert "redo integer8 for laplace and helm"
This reverts commit 99c9bbd495082fd897cf51779137e2acbc474a25.

+++ src/Helmholtz/h3dcommon.f
@@ -168,7 +168,7 @@ c     add mpole to mpole2
 c
 c----------------------------------------------------------------------
       implicit real *8 (a-h,o-z)
-      integer(8) nd
+      integer nd
       complex *16 mpole(nd,0:nterms,-nterms:nterms)
       complex *16 mpole2(nd,0:nterms,-nterms:nterms)
 c
