+++ test/Laplace/test_lfmm3d.f
@@ -1,5 +1,5 @@
       implicit none
-      integer(8) ns,nt
+      integer ns,nt
       double precision, allocatable :: source(:,:),targ(:,:)
       double precision, allocatable :: charge(:)
       double precision, allocatable :: dipvec(:,:)
@@ -9,10 +9,10 @@
 
       double precision eps
       double complex eye
-      integer(8) i,j,k,ntest,ier
-      integer(8) ifcharge,ifdipole,ifpgh,ifpghtarg
+      integer i,j,k,ntest,ier
+      integer ifcharge,ifdipole,ifpgh,ifpghtarg
       double precision err,hkrand
-      integer(8) ipass(27),len1,ntests,isum
+      integer ipass(27),len1,ntests,isum
       character(len=100) str1
       
 
@@ -981,7 +981,7 @@ c
 
       implicit none
       double complex zk
-      integer(8) ns,nt,ifcharge,ifdipole,ifpgh,ifpghtarg
+      integer ns,nt,ifcharge,ifdipole,ifpgh,ifpghtarg
       
       double precision source(3,*),targ(3,*)
       double precision dipvec(3,*)
@@ -990,7 +990,7 @@ c
       double precision pot(*),pottarg(*),grad(3,*),gradtarg(3,*)
       double precision hess(6,*),hesstarg(6,*)
 
-      integer(8) i,j,ntest,nd,l
+      integer i,j,ntest,nd,l
 
       double precision err,ra
